% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{all.equal,matrixORbigmatrix,matrixORbigmatrix-method}
\alias{all.equal,matrixORbigmatrix,matrixORbigmatrix-method}
\title{all.equal}
\usage{
\S4method{all.equal}{matrixORbigmatrix,matrixORbigmatrix}(target, current,
  tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{target}{A \code{matrix} or \code{big.matrix} object}

\item{current}{A \code{matrix} or \code{big.matrix} object}

\item{tolerance}{tolerance for rounding error}
}
\value{
Either \code{TRUE} or \code{FALSE} indicating if all elements are
equivlanet.
}
\description{
all.equal method for comparing elements between big.matrix and
matrix objects.
}

