% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{crossprod,big.matrix,missing-method}
\alias{crossprod,big.matrix}
\alias{crossprod,big.matrix,big.matrix-method}
\alias{crossprod,big.matrix,missing-method}
\alias{tcrossprod,big.matrix,big.matrix-method}
\alias{tcrossprod,big.matrix,missing-method}
\title{big.matrix crossproduct}
\usage{
\S4method{crossprod}{big.matrix,missing}(x, y)

\S4method{crossprod}{big.matrix,big.matrix}(x, y)

\S4method{tcrossprod}{big.matrix,missing}(x, y)

\S4method{tcrossprod}{big.matrix,big.matrix}(x, y)
}
\arguments{
\item{x}{A big.matrix}

\item{y}{A big.matrix}
}
\value{
A big.matrix
}
\description{
Return the matrix cross-product of two conformable
big.matrix objects.  This is equivalent to t(x) %*% y (crossprod)
or x %*% t(t) (tcrossprod) but faster as no data transfer between
device and host is required.
}
\author{
Charles Determan Jr.
}

