% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_extras.R
\name{ddot}
\alias{ddot}
\title{Dot product of two vectors}
\usage{
ddot(N = NULL, X, INCX = 1L, Y, INCY = 1L)
}
\arguments{
\item{N}{Optional integer giving the number of elements. Defaults to `length(X)`.}

\item{X}{Double-precision vector, matrix or [`bigmemory::big.matrix`] input.}

\item{INCX}{Integer stride for traversing `X`.}

\item{Y}{Double-precision object updated in place.}

\item{INCY}{Integer stride for traversing `Y`.}
}
\value{
Numeric scalar containing the dot product.
}
\description{
Dot product of two vectors
}
\examples{
ddot(X = 1:3, Y = c(2, 4, 6))

}
