% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_extras.R
\name{dnrm2}
\alias{dnrm2}
\title{Euclidean norm (2-norm)}
\usage{
dnrm2(N = NULL, X, INCX = 1L)
}
\arguments{
\item{N}{Optional integer giving the number of elements. Defaults to `length(X)`.}

\item{X}{Double-precision vector, matrix or [`bigmemory::big.matrix`] input.}

\item{INCX}{Integer stride for traversing `X`.}
}
\value{
Numeric scalar containing the Euclidean norm.
}
\description{
Euclidean norm (2-norm)
}
