% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_extras.R
\name{dsum}
\alias{dsum}
\title{Sum of elements}
\usage{
dsum(N = NULL, X, INCX = 1L)
}
\arguments{
\item{N}{Optional integer giving the number of elements. Defaults to `length(X)`.}

\item{X}{Double-precision vector, matrix or [`bigmemory::big.matrix`] input.}

\item{INCX}{Integer stride for traversing `X`.}
}
\value{
Numeric scalar giving the sum of elements of `X`.
}
\description{
Sum of elements
}
