% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_extras.R
\name{dxyz}
\alias{dxyz}
\title{Three-dimensional cross product}
\usage{
dxyz(X, Y, Z)
}
\arguments{
\item{X}{Numeric vector of length three, matrix with three rows, or big.matrix.}

\item{Y}{Numeric object matching the shape of `X`.}

\item{Z}{Optional output container.}
}
\value{
The updated object `Z` containing the cross product.
}
\description{
Three-dimensional cross product
}
\examples{
dxyz(X = c(1, 0, 0), Y = c(0, 1, 0))

}
