% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_extras.R
\name{idmin}
\alias{idmin}
\title{Index of the minimum element}
\usage{
idmin(N = NULL, X, INCX = 1L)
}
\arguments{
\item{N}{Optional integer giving the number of elements. Defaults to `length(X)`.}

\item{X}{Double-precision vector, matrix or [`bigmemory::big.matrix`] input.}

\item{INCX}{Integer stride for traversing `X`.}
}
\value{
Integer index (1-based) of the smallest entry in `X`.
}
\description{
Index of the minimum element
}
