% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t_inplace_wrapper}
\alias{t_inplace_wrapper}
\title{Big Matrix In Place Transposition}
\usage{
t_inplace_wrapper(X, LOW_MEM)
}
\arguments{
\item{X}{big.matrix object}

\item{LOW_MEM}{boolean option to choose slower, low memory option}
}
\description{
Big Matrix In Place Transposition
}

