% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t_wrapper}
\alias{t_wrapper}
\title{Big Matrix Transposition}
\usage{
t_wrapper(X, Y)
}
\arguments{
\item{X}{big.matrix object}

\item{Y}{big.matrix object (to be filled)}

\item{LOW_MEM}{boolean option to choose slower, low memory option}
}
\description{
Big Matrix Transposition
}

