\name{bigalgebra-package}
\Rdversion{1.1}
\alias{bigalgebra-package}
\alias{bigalgebra}
\docType{package}
\title{
This package provides blas dgemm functionality for native R matrices and
big.matrix objects.
}
\description{
See title
}
\details{
\tabular{ll}{
Package: \tab bigalgebra\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2010-04-08\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
Use this package for multiplying matrices using a BLAS library.
}
\author{
Michael Kane

Maintainer: Mike <kaneplusplus@gmail.com>
}
\references{
http://www.netlib.org/blas/dgemm.f
}
\keyword{ package }
\seealso{
\code{\link[bigmemory-package]{big.matrix}} 
}
\examples{
print('bigalgebra')
}
